backend/pyproject.toml
[project]
name = "voice-cloning-demo"
version = "0.1.0"
description = "LiveKit demo for voice cloning with Resemble AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "langfuse==2.60.8",
    "livekit-agents[deepgram,resemble,openai,silero,turn-detector]",
    "livekit-api",
    "python-dotenv",
    "uvicorn",
]